+++ js/bitjs/v1.2.6/archive/rarvm.js
@@ -6,6 +6,8 @@
  * Copyright(c) 2017 Google Inc.
  */
 
+import { BitStream } from '../io/bitstream.js';
+
 /**
  * CRC Implementation.
  */
@@ -57,7 +59,7 @@ function InitCRC() {
 /**
  * @param {number} startCRC
  * @param {Uint8Array} arr
- * @return {number}
+ * @returns {number}
  */
 function CRC(startCRC, arr) {
   if (CRCTab[1] == 0) {
@@ -104,13 +106,13 @@ function CRC(startCRC, arr) {
 /**
  * RarVM Implementation.
  */
-const VM_MEMSIZE = 0x40000;
-const VM_MEMMASK = (VM_MEMSIZE - 1);
-const VM_GLOBALMEMADDR = 0x3C000;
-const VM_GLOBALMEMSIZE = 0x2000;
-const VM_FIXEDGLOBALSIZE = 64;
-const MAXWINSIZE = 0x400000;
-const MAXWINMASK = (MAXWINSIZE - 1);
+export const VM_MEMSIZE = 0x40000;
+export const VM_MEMMASK = (VM_MEMSIZE - 1);
+export const VM_GLOBALMEMADDR = 0x3C000;
+export const VM_GLOBALMEMSIZE = 0x2000;
+export const VM_FIXEDGLOBALSIZE = 64;
+export const MAXWINSIZE = 0x400000;
+export const MAXWINMASK = (MAXWINSIZE - 1);
 
 /**
  */
@@ -204,7 +206,7 @@ const VM_OpType = {
  * variables that use the above enums.
  * @param {Object} obj
  * @param {number} val
- * @return {string} The key/enum value as a string.
+ * @returns {string} The key/enum value as a string.
  */
 function findKeyForValue(obj, val) {
   for (let key in obj) {
@@ -248,7 +250,7 @@ class VM_PreparedOperand {
     this.Addr = null;
   };
 
-  /** @return {string} */
+  /** @returns {string} */
   toString() {
     if (this.Type === null) {
       return 'Error: Type was null in VM_PreparedOperand';
@@ -278,7 +280,7 @@ class VM_PreparedCommand {
     this.Op2 = new VM_PreparedOperand();
   }
 
-  /** @return {string} */
+  /** @returns {string} */
   toString(indent) {
     if (this.OpCode === null) {
       return 'Error: OpCode was null in VM_PreparedCommand';
@@ -319,7 +321,7 @@ class VM_PreparedProgram {
     this.FilteredData = null;
   }
 
-  /** @return {string} */
+  /** @returns {string} */
   toString() {
     let s = '{\n  Cmd: [\n';
     for (let i = 0; i < this.Cmd.length; ++i) {
@@ -334,7 +336,7 @@ class VM_PreparedProgram {
 
 /**
  */
-class UnpackFilter {
+export class UnpackFilter {
   constructor() {
     /** @type {number} */
     this.BlockStart = 0;
@@ -448,7 +450,7 @@ const StdList = [
 /**
  * @constructor
  */
-class RarVM {
+export class RarVM {
   constructor() {
     /** @private {Uint8Array} */
     this.mem_ = null;
@@ -471,7 +473,7 @@ class RarVM {
 
   /**
    * @param {Uint8Array} code
-   * @return {VM_StandardFilters}
+   * @returns {VM_StandardFilters}
    */
   isStandardFilter(code) {
     const codeCRC = (CRC(0xffffffff, code, code.length) ^ 0xffffffff) >>> 0;
@@ -486,7 +488,7 @@ class RarVM {
   /**
    * @param {VM_PreparedOperand} op
    * @param {boolean} byteMode
-   * @param {bitjs.io.BitStream} bstream A rtl bit stream.
+   * @param {BitStream} bstream A rtl bit stream.
    */
   decodeArg(op, byteMode, bstream) {
     const data = bstream.peekBits(16);
@@ -574,7 +576,7 @@ class RarVM {
 
   /**
    * @param {Array<VM_PreparedCommand>} preparedCodes
-   * @return {boolean}
+   * @returns {boolean}
    */
   executeCode(preparedCodes) {
     let codeIndex = 0;
@@ -808,7 +810,7 @@ class RarVM {
 
     //InitBitInput();
     //memcpy(InBuf,Code,Min(CodeSize,BitInput::MAX_SIZE));
-    const bstream = new bitjs.io.BitStream(code.buffer, true /* rtl */);
+    const bstream = new BitStream(code.buffer, true /* rtl */);
 
     // Calculate the single byte XOR checksum to check validity of VM code.
     let xorSum = 0;
@@ -970,8 +972,8 @@ class RarVM {
   /**
    * Static function that reads in the next set of bits for the VM
    * (might return 4, 8, 16 or 32 bits).
-   * @param {bitjs.io.BitStream} bstream A RTL bit stream.
-   * @return {number} The value of the bits read.
+   * @param {BitStream} bstream A RTL bit stream.
+   * @returns {number} The value of the bits read.
    */
   static readData(bstream) {
     // Read in the first 2 bits.