+++ js/cbrjs/bitjs.js
@@ -1,0 +1,5 @@
+import { getUnarchiver, UnarchiveEventType } from '../bitjs/v1.2.6/archive/decompress.js';
+
+window.CBRJS = window.CBRJS || {};
+window.CBRJS.getUnarchiver = getUnarchiver;
+window.CBRJS.UnarchiveEventType = UnarchiveEventType;